+++ resources/views/judgement/index.blade.php
@@ -12,7 +12,13 @@
                         <a> 此評分系統根據路況、天數及所需體力來量化，並依據分數高低來訂定隊伍難度等級。</a>
                     </div>
                 </div>
-                <livewire:judgement-component.form/>
+
+                <!-- 評分表單 -->
+                <livewire:judgement-component.form :judgement="null" :mode="'create'"/>
+
+                <!-- 評分紀錄表 -->
+                <livewire:judgement-component.page/>
+
                 <!-- 難度分級表 -->
                 <div class="row gy-4 justify-content-center text-center">
                     <h1>難度分級表</h1>
@@ -86,7 +92,6 @@
                         </table>
                     </div>
                 </div><!-- 難度分級表 End -->
-
 
                 <!-- 路況分級表 -->
                 <div class="row gy-4 justify-content-center text-center">
@@ -205,86 +210,6 @@
                         </table>
                     </div>
                 </div><!-- 難度分級表 End  -->
-
-                <!-- 評分紀錄表 -->
-                <div class="row gy-4 justify-content-center text-center">
-                    <h1>評分紀錄表</h1>
-                    <a>過去隊伍難度的評分紀錄。</a>
-
-                    <div class="col-md-12 text-center mb-5 table-responsive">
-                        <table class="table table-light table-bordered table-striped">
-                            <thead>
-                            <tr>
-                                <th scope="col">#</th>
-                                <th scope="col">路線名稱</th>
-                                <th scope="col">總天數</th>
-                                <th scope="col">傳統路</th>
-                                <th scope="col">非傳統路</th>
-                                <th scope="col">路線</th>
-                                <th scope="col">路標</th>
-                                <th scope="col">地形</th>
-                                <th scope="col">植被</th>
-                                <th scope="col">體力</th>
-                                <th scope="col">背水天數</th>
-                                <th scope="col">難度總分</th>
-                                <th scope="col">隊伍難度</th>
-                                <!-- 幹部才能編輯 -->
-                                @auth
-                                    @if(Auth::user()->role > 0)
-                                        <th scope="col">編輯/刪除</th>
-                                    @endif
-                                @endauth
-                            </tr>
-                            </thead>
-                            <tbody>
-                            @if (!$judgements->count())
-                                <tr>
-                                    <td colspan= {{ $judgements_column_number }}>目前暫無評分紀錄</td>
-                                </tr>
-                            @else
-                                @foreach($judgements as $judgement)
-                                    <tr>
-                                        <td>{{ $loop->index + 1 }}</td>
-                                        @if ($judgement->trip_tag == 0)
-                                            <td>{{ $judgement->name }}</td>
-                                        @elseif ($judgement->trip_tag == 1)
-                                            <td>{{ $judgement->name }} <span style="color:red">(壓縮行程)</span></td>
-                                        @elseif ($judgement->trip_tag == 2)
-                                            <td>{{ $judgement->name }} <span style="color:green">(寬鬆行程)</span></td>
-                                        @endif
-                                        <td>{{ $judgement->normal_day + $judgement->abnormal_day}}</td>
-                                        <td>{{ $judgement->normal_day }} 天</td>
-                                        <td>{{ $judgement->abnormal_day }} 天</td>
-                                        <td>{{ $level_array[$judgement->level] }}</td>
-                                        <td>{{ $judgement->road }} 分</td>
-                                        <td>{{ $judgement->terrain }} 分</td>
-                                        <td>{{ $judgement->plant }} 分</td>
-                                        <td>{{ $judgement->energy }}</td>
-                                        <td>{{ $judgement->water }} 天</td>
-                                        <td>{{ $judgement->score }} 分</td>
-                                        <td>{{ $judgement->result_level }}</td>
-                                        <!-- 幹部才能編輯 -->
-                                        @auth
-                                            @if(Auth::user()->role > 0)
-                                                <td>
-                                                    <form action="{{ route('judgement.destroy', $judgement->id) }}"
-                                                          method="POST">
-                                                        <button type="button" class="bi bi-pencil-square"
-                                                                onclick="window.location='{{ route('judgement.edit', $judgement->id) }}'"></button>
-                                                        @csrf
-                                                        @method('DELETE')
-                                                        <button type="submit" class="bi bi-trash"></button>
-                                                    </form>
-                                                </td>
-                                            @endif
-                                        @endauth
-                                    </tr>
-                                @endforeach
-                            @endif
-                            </tbody>
-                        </table>
-                    </div>
-                </div><!-- 評分紀錄表 End  -->
 
             </div>
         </section>
